nixosConfigurations/universe/skinprox: remove skins.tw
While skins.tw (or rather teedata.net) is a popular source for
Teeworlds/DDNet skins, it has an inherently broken API that returns the
default teeworlds skin when querying for an unknown skin. A bitter
side-effect of this is that users' data directory is going to get filled
up by copies of the default skin, every time an unknown skin is
requested.

The developers of teedata.net have not responded to my inquiries about
this.

Signed-off-by: Sefa Eyeoglu <contact@scrumplex.net>